// client/src/types/category.type.ts
// src/types/category.ts

export type CategoryCode =
  | "GRO"
  | "HIM"
  | "PT"
  | "HR"
  | "UB"
  | "HLT"
  | "EDU"
  | "CLT"
  | "ENT"
  | "TRV"
  | "SPT"
  | "BTY"
  | "GFT"
  | "REP"
  | "HOB"
  | "PTS"
  | "INS"
  | "TX"
  | "FUE"
  | "INV"
  | "CHY"
  | "IC"
  | "CG"
  | "VD"
  | "SUB"
  | "CRE";